Move the function to the proper place
